feat: PlaceService 카카오 panel3 기반으로 변경

+++ src/main/java/com/matzip/common/dto/LocationDto.java
@@ -1,18 +1,16 @@
 package com.matzip.common.dto;
 
-import lombok.Builder;
-import lombok.Getter;
+import lombok.*;
 
-import java.math.BigDecimal;
 
 @Getter
 @Builder
 public class LocationDto {
 
-    private BigDecimal latitude;
-    private BigDecimal longitude;
+    private double latitude;
+    private double longitude;
 
-    public static LocationDto of(BigDecimal latitude, BigDecimal longitude) {
+    public static LocationDto of(double latitude, double longitude) {
         return LocationDto.builder()
                 .latitude(latitude)
                 .longitude(longitude)
